+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## unreleased
 * Local Payment Methods
-  * Add `bic` to `BTLocalPaymentRequest`
+  * Add `bic` (Bank Identification Code) to `BTLocalPaymentRequest`
 
 ## 5.0.1 (2021-03-01)
 * SPM