+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## unreleased
 * Local Payment Methods
   * Add `bic` (Bank Identification Code) to `BTLocalPaymentRequest`
+* Apple Pay
+  * Add support for `PKPaymentNetworkElo` to Apple Pay configuration
 
 ## 5.0.1 (2021-03-01)
 * SPM